Extract PHP code emitter from base class
