+++ src/main/php/lang/ast/emit/PHP71.class.php
@@ -1,13 +1,11 @@
 <?php namespace lang\ast\emit;
 
-use lang\ast\Emitter;
-
 /**
  * PHP 7.1 syntax
  *
  * @see  https://wiki.php.net/rfc#php_71
  */
-class PHP71 extends Emitter {
+class PHP71 extends PHP {
   use OmitPropertyTypes;
   use RewriteNullCoalesceAssignment, RewriteLambdaExpressions;
