+++ src/main/php/lang/ast/emit/PHP72.class.php
@@ -1,13 +1,11 @@
 <?php namespace lang\ast\emit;
 
-use lang\ast\Emitter;
-
 /**
  * PHP 7.2 syntax
  *
  * @see  https://wiki.php.net/rfc#php_72
  */
-class PHP72 extends Emitter {
+class PHP72 extends PHP {
   use OmitPropertyTypes;
   use RewriteNullCoalesceAssignment, RewriteLambdaExpressions;
