+++ src/main/php/lang/ast/emit/PHP74.class.php
@@ -1,13 +1,11 @@
 <?php namespace lang\ast\emit;
 
-use lang\ast\Emitter;
-
 /**
  * PHP 7.4 syntax
  *
  * @see  https://wiki.php.net/rfc#php_74
  */
-class PHP74 extends Emitter {
+class PHP74 extends PHP {
   use RewriteBlockLambdaExpressions;
 
   protected $unsupported= [
