+++ src/main/php/lang/ast/emit/PHP80.class.php
@@ -1,13 +1,11 @@
 <?php namespace lang\ast\emit;
 
-use lang\ast\Emitter;
-
 /**
  * PHP 8.0 syntax
  *
  * @see  https://wiki.php.net/rfc#php_80
  */
-class PHP80 extends Emitter {
+class PHP80 extends PHP {
   use RewriteBlockLambdaExpressions;
 
   protected $unsupported= [
